# Fix very slow iteration over single read