@lbergelson 2017-11-08
@PolinaBevad Actually, on further thought I think this might introduce a performance regression for non-gapped reads.  Previously, if the read wasn't gapped it would short circuit as soon as it reached a point where x > locusToStop.

Using a slight modification of your test code, I measured the following:

Cigar = `36000M`

| original code | removeIf |
| --- | --- |
| 0.206 sec |  11.741 sec |

I think we may need a solution that does a fast remove and short circuits.  I don't see a neat way to do that in a one liner.