@PolinaBevad 2017-11-09
@lbergelson Thank you for review!
Yes, I have tested this case and now see the problem. Code from new commit solve it: measurement of test for old and new implementations with cigar 36000M without gaps:

original, sec| new, sec
-----|------
0.941 | 0.92